+++ app/controllers/application_controller.rb
@@ -14,7 +14,7 @@ class ApplicationController < ActionController::Base
     if current_user
       render file: 'public/403.html', status: :forbidden, layout: false
     else
-      redirect_to new_user_session_path
+      redirect_to new_user_session_path(redirect_to: request.path)
     end
   end
 
@@ -22,4 +22,14 @@ class ApplicationController < ActionController::Base
     @user ||= UserDecorator.new(current_user)
   end
   helper_method :user
+
+  def after_sign_in_path_for(resource)
+    origin = request.env['omniauth.origin']
+    if origin
+      query = CGI.unescape(URI.parse(origin).query)
+      path = CGI::parse(query)['redirect_to']&.first
+    end
+
+    path || super(resource)
+  end
 end