+++ app/serializers/resource_index_serializer.rb
@@ -1,0 +1,6 @@
+class ResourceIndexSerializer < ActiveModel::Serializer
+  attributes :id, :name, :has_invalid_mocks?
+  attributes :request_route_ids
+  attributes :response_ids
+  has_many :used_resources, serializer: IdSerializer
+end