+++ db/migrate/20180219094508_add_ignore_ssl_to_proxy_configuration.rb
@@ -1,0 +1,5 @@
+class AddIgnoreSslToProxyConfiguration < ActiveRecord::Migration[5.1]
+  def change
+    add_column :proxy_configurations, :ignore_ssl, :boolean, default: false, null: false
+  end
+end