+++ test/controllers/projects_controller_test.rb
@@ -20,7 +20,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
   test "should not show project (not authenticated)" do
     sign_out :user
     get project_path(@project)
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should get new" do
@@ -31,7 +31,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
   test "should not get new (not authenticated)" do
     sign_out :user
     get new_project_path
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should get edit" do
@@ -42,7 +42,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
   test "should not get edit (not authenticated)" do
     sign_out :user
     get edit_project_path(@project)
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should create project" do
@@ -64,7 +64,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
     assert_no_difference('Project.count') do
       post projects_path, params: { project: { description: 'My description', title: 'My Project' }}
     end
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should update project" do
@@ -72,6 +72,32 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
     assert_redirected_to project_path(@project)
     @project.reload
     assert_equal 'New Description', @project.description
+  end
+
+  test "update project proxy configuration" do
+    put project_path(@project), params: {
+      project: {
+        proxy_configuration_attributes: {
+          target_base_url: 'https://api.com'
+        }
+      }
+    }
+
+    assert_equal 'https://api.com', @project.reload.proxy_configuration.target_base_url
+  end
+
+  test "update project delete proxy configuration if no target_base_url" do
+    @project.create_proxy_configuration(target_base_url: 'https://api.com')
+
+    assert_difference 'ProxyConfiguration.count', -1 do
+      put project_path(@project), params: {
+        project: {
+          proxy_configuration_attributes: {
+            target_base_url: ''
+          }
+        }
+      }
+    end
   end
 
   test "should not update project" do
@@ -88,7 +114,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
     put project_path(@project), params: { project: { title: "New title" }}
     @project.reload
     assert_equal project_original_title, @project.title
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should delete project" do
@@ -103,7 +129,7 @@ class ProjectsControllerTest < ControllerWithAuthenticationTest
     assert_no_difference 'Project.count' do
       delete project_path(@project)
     end
-    assert_redirected_to new_user_session_path
+    assert_redirected_to new_user_session_path(redirect_to: request.path)
   end
 
   test "should get zip file with all json schemas" do