+++ test/factories/proxy_configurations.rb
@@ -1,0 +1,7 @@
+FactoryGirl.define do
+
+  factory :proxy_configuration do
+    project
+    target_base_url 'https://api-develop.herokuapp.com'
+  end
+end